Refactor into contact person and person roles components
